pyproject.toml
[project]
name = "rag-knowledge-mcp"
version = "0.1.0"
description = "RAG Knowledge Base MCP Server"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.3.0",
    "numpy<2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]
